creados modulos -Jugador -Equipo -Evento -Main -InicioApp
