--- app/src/main/java/com/example/dannyang27/sportpoints/activities/ConfiguracionPerfil.java
@@ -1,15 +1,0 @@
-package com.example.dannyang27.sportpoints.activities;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-import com.example.dannyang27.sportpoints.R;
-
-public class ConfiguracionPerfil extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.configuracion_perfil);
-    }
-}
