+++ app/src/main/java/com/example/dannyang27/sportpoints/activities/OpcionesChoser.java
@@ -9,6 +9,8 @@ import android.view.View;
 import android.widget.Button;
 
 import com.example.dannyang27.sportpoints.R;
+import com.example.dannyang27.sportpoints.activities.equipo.EquipoInfo;
+import com.example.dannyang27.sportpoints.activities.evento.EventoInfo;
 
 public class OpcionesChoser extends AppCompatActivity {
