+++ app/src/main/java/com/example/dannyang27/sportpoints/activities/evento/EventoInfo.java
@@ -1,15 +1,15 @@
-package com.example.dannyang27.sportpoints.activities;
+package com.example.dannyang27.sportpoints.activities.evento;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.example.dannyang27.sportpoints.R;
 
-public class EquipoInfo extends AppCompatActivity {
+public class EventoInfo extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.equipo_info);
+        setContentView(R.layout.evento_info);
     }
 }
