+++ app/src/main/java/com/example/dannyang27/sportpoints/activities/inicioApp/PerfilAlta.java
@@ -1,4 +1,4 @@
-package com.example.dannyang27.sportpoints.activities;
+package com.example.dannyang27.sportpoints.activities.inicioApp;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -9,6 +9,8 @@ import android.widget.EditText;
 import android.widget.ImageView;
 
 import com.example.dannyang27.sportpoints.R;
+import com.example.dannyang27.sportpoints.activities.Deportes;
+import com.example.dannyang27.sportpoints.activities.main.MainActivity;
 
 public class PerfilAlta extends AppCompatActivity {
