+++ app/src/main/java/com/example/dannyang27/sportpoints/activities/jugador/ConfiguracionPerfil.java
@@ -1,15 +1,15 @@
-package com.example.dannyang27.sportpoints.activities;
+package com.example.dannyang27.sportpoints.activities.jugador;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
 import com.example.dannyang27.sportpoints.R;
 
-public class EventoInfo extends AppCompatActivity {
+public class configuracionPerfil extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.evento_info);
+        setContentView(R.layout.configuracion_perfil);
     }
 }
