+++ app/src/main/java/com/example/dannyang27/sportpoints/activities/main/MainActivity.java
@@ -1,14 +1,15 @@
-package com.example.dannyang27.sportpoints.activities;
+package com.example.dannyang27.sportpoints.activities.main;
 
 import android.content.Intent;
 import android.graphics.Typeface;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.TextView;
 
 import com.example.dannyang27.sportpoints.R;
+import com.example.dannyang27.sportpoints.activities.OpcionesChoser;
+import com.example.dannyang27.sportpoints.activities.inicioApp.PerfilAlta;
 import com.google.firebase.database.*;
 
 public class MainActivity extends AppCompatActivity {
