+++ app/src/main/res/layout/configuracion_perfil.xml
@@ -8,6 +8,6 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.dannyang27.sportpoints.activities.ConfiguracionPerfil">
+    tools:context="com.example.dannyang27.sportpoints.activities.jugador.ConfiguracionPerfil">
 
 </RelativeLayout>
