+++ app/src/main/res/layout/equipo_info.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.dannyang27.sportpoints.activities.EquipoInfo">
+    tools:context="com.example.dannyang27.sportpoints.activities.equipo.EquipoInfo">
 
     <SearchView
         android:layout_width="250dp"
