+++ app/src/main/res/layout/login.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.dannyang27.sportpoints.activities.Login">
+    tools:context="com.example.dannyang27.sportpoints.activities.inicioApp.Login">
 
     <TextView
         android:text="Introduzca el email"
