+++ app/src/main/res/layout/perfil_alta.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.dannyang27.sportpoints.activities.PerfilAlta">
+    tools:context="com.example.dannyang27.sportpoints.activities.inicioApp.PerfilAlta">
 
     <ImageView
         android:layout_width="100dp"
