+++ app/src/main/res/layout/principal.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.dannyang27.sportpoints.activities.Principal">
+    tools:context="com.example.dannyang27.sportpoints.activities.Deportes$Principal">
 
     <ListView
         android:layout_width="match_parent"
