+++ app/src/main/res/layout/tipo_usuario.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.dannyang27.sportpoints.activities.TipoUsuario">
+    tools:context="com.example.dannyang27.sportpoints.activities.inicioApp.TipoUsuario">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
